requirements.txt
# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0

# Machine Learning
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning
tensorflow>=2.15.0

# Database - POSTGRESQL/SUPABASE
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0